+++ NEWS.md
@@ -9,6 +9,8 @@
 - Move TIMEVAL to its own column in variables1, instead of having it as part of
 `variable-type`. (#265)
 - Rename 'pxsave' to 'px_save'. (#273)
+- Rename 'figures' to 'px_figures', 'order' to 'px_order', 'add_totals' to 
+'px_add_totals'. (#274)
 
 ## New features
 - Add modifying functions: `map()`, `baseperiod()`, `domain()`, `elimination()`,